Turn YBD and KBAS into proper Python packages
This means you can now 'import ybd' from anywhere in your system and use
its functionality in other programs.

Using special __main__.py modules, you can also now do 'python -m ybd' to
execute YBD, and same for KBAS. The top-level kbas.py and ybd.py scripts
are now just wrappers, that effectively do the same as 'python -m ybd'
and 'python -m kbas'. See https://www.python.org/dev/peps/pep-0338/ for
way more information about that approach than you ever wanted.
