+++ kbas/kbas.tpl
