+++ ybd/__init__.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 # Copyright (C) 2015  Codethink Limited
 #
 # This program is free software; you can redistribute it and/or modify
@@ -13,6 +12,14 @@
 # You should have received a copy of the GNU General Public License along
 # with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import runpy
 
-runpy.run_module('ybd')
+import app
+import assembly
+import cache
+import defaults
+import definitions
+import deployment
+import repos
+import sandbox
+import utils
+import wrangler
